fix plotting issues
